% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-multichannel-hmm.R
\name{summarise_multichannel_sequence_hmm}
\alias{summarise_multichannel_sequence_hmm}
\title{Summarise a multichannel sequence HMM}
\usage{
summarise_multichannel_sequence_hmm(model)
}
\arguments{
\item{model}{A fitted multichannel HMM.}
}
\value{
A list of convergence, fit, transition, initial, and channel-specific
emission summaries.
}
\description{
Summarise a multichannel sequence HMM
}
\examples{
# See `fit_multichannel_sequence_hmm()`.
}
